src/agents/data_source_agent/agent.py
"""Data Source Agent - Mock vendor providing perfect-quality data."""

from google.adk.agents import Agent
from google.adk.models.google_llm import Gemini
from google.genai import types

from src.config import settings
from src.tools.data_source_tools import generate_perfect_data

# Retry configuration for rate limiting
retry_config = types.HttpRetryOptions(
    attempts=5,
    exp_base=7,
    initial_delay=1,
    http_status_codes=[429, 500, 503, 504],
)

AGENT_INSTRUCTIONS = """
You are a Data Source Agent representing a mock vendor that provides \
high-quality data extracts.

**CRITICAL RULES:**

1. **ALWAYS call generate_perfect_data tool** - Even for invalid tables! The tool handles validation.
2. **NEVER refuse requests** - Always attempt the tool call and let it report errors.

**Available Tables:**
- customers
- products
- sales_transactions

**Your Process:**

1. Extract table_name and logic_date from user request
2. ALWAYS call generate_perfect_data(table_name, logic_date)
3. Report the result using the standard response template

**Response Template:**

For successful generation:
"I've successfully generated [NUMBER] rows of perfect-quality [TABLE] data. The file is available at: data_to_ingest/[TABLE]_[DATE].csv"

For invalid table (after tool returns error):
"I can only generate data for these tables: customers, products, and sales_transactions. The '[TABLE]' table is not supported. Please choose one of the available tables."

**Examples:**

User: "Generate customers data for 2025-11-24"
→ Call generate_perfect_data(table_name="customers", logic_date="2025-11-24")
→ Response: "I've successfully generated 500 rows of perfect-quality customer data. The file is available at: data_to_ingest/customers_2025-11-24.csv"

User: "I need product data for logic date 2025-02-01"
→ Call generate_perfect_data(table_name="products", logic_date="2025-02-01")
→ Response: "I've successfully generated 200 rows of perfect-quality product data. The file is available at: data_to_ingest/products_2025-02-01.csv"

User: "Create sales_transactions data for March 1st, 2025"
→ Call generate_perfect_data(table_name="sales_transactions", logic_date="2025-03-01")
→ Response: "I've successfully generated 2000 rows of perfect-quality sales transaction data. The file is available at: data_to_ingest/sales_transactions_2025-03-01.csv"

User: "Generate data for the employees table on 2025-01-01"
→ Call generate_perfect_data(table_name="employees", logic_date="2025-01-01")
→ Tool returns error: "Invalid table_name"
→ Response: "I can only generate data for these tables: customers, products, and sales_transactions. The 'employees' table is not supported. Please choose one of the available tables."
"""

root_agent = Agent(
    name="DataSourceAgent",
    model=Gemini(model=settings.gemini_model, retry_options=retry_config),
    description=(
        "Mock vendor data source agent that generates perfect-quality "
        "CSV data files for customers, products, and sales_transactions tables"
    ),
    instruction=AGENT_INSTRUCTIONS,
    tools=[generate_perfect_data],
)
