pyproject.toml
[tool.poetry]
name = "kaggle-google-ai-agents"
version = "0.1.0"
description = "Data Engineer Assistant Agent - AI-powered assistant for data quality, pipeline management, and analytics"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
# Agent Framework
google-generativeai = "^0.8.0"

# Data Processing & Storage
duckdb = "^1.1.0"

# Chat Interface

# Visualization
plotly = "^5.17.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
tabulate = "^0.9.0"

# Utilities
pandas = "^2.1.0"
numpy = "^1.24.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
rich = "^13.8.0"
typer = "^0.12.0"
loguru = "^0.7.2"

# Jupyter Support
jupyter = "^1.0.0"
notebook = "^7.0.0"
ipykernel = "^6.25.0"
ipython = "^8.15.0"
huggingface-hub = ">=0.26.0"
polars = ">=1.35.0"
google-adk = {extras = ["eval"], version = "^1.19.0"}
faker = "^38.2.0"
aiosqlite = "^0.21.0"
a2a-sdk = "^0.3.19"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
explicit_package_bases = true
namespace_packages = true
mypy_path = "$MYPY_CONFIG_FILE_DIR"

[[tool.mypy.overrides]]
module = "src.*"
ignore_errors = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
